Exlm 5192 - Add header allowed domain inside the exl-config (#2779)
* docs: add design spec for header v2 domain-allowlist rollout

Captures the plan to replace the isHeaderV2 feature flag with a
headerv2allowedDomains config check, sharing config-fetch logic with
Premium Learning's existing domain-gating pattern.

* refactor: extract shared exlm-config-utils from premium-learning-utils

Moves the private getExlmConfig/fetchExlmConfig helpers into a shared
module and adds isDomainAllowed(), so the comma-separated domain-list
check isn't duplicated. No behavior change for Premium Learning.

* feat: gate header v2 by headerv2allowedDomains instead of feature flag

Replaces the isFeatureEnabled('isHeaderV2') page-metadata check with a
domain-allowlist check via the shared exlm-config-utils, matching the
rollout pattern Premium Learning already uses. Verified in a browser
that v1 renders under the current (empty) config, and v2 renders when
the current hostname is in headerv2allowedDomains.

* docs: record accepted latency tradeoff in header v2 spec

Self-review flagged that the domain check blocks header rendering on
a network fetch (unlike the fire-and-forget pattern Premium Learning
uses for the same config). Decision: accept as-is since the config
fetch is small and edge-cached; revisit if render latency regresses.

* remove spec file

* fix feature-flag update
