+++ blocks/header/header.js
@@ -16,6 +16,8 @@ import {
 import getProducts from '../../scripts/utils/product-utils.js';
 import { isSignedInUser } from '../../scripts/auth/profile.js';
 import { isPLEligible } from '../../scripts/utils/premium-learning-utils.js';
+import { isDomainAllowed } from '../../scripts/utils/exlm-config-utils.js';
+import isFeatureEnabled from '../../scripts/utils/feature-flag-utils.js';
 import {
   decoratorState,
   isMobile,
@@ -29,7 +31,6 @@ import {
 import { decorateIcons, getMetadata } from '../../scripts/lib-franklin.js';
 import LanguageBlock from '../language/language.js';
 import ProfileMenu from './profile-menu.js';
-import isFeatureEnabled from '../../scripts/utils/feature-flag-utils.js';
 
 /**
  *  @typedef {Object} CommunityOptions
@@ -818,7 +819,7 @@ customElements.define('exl-header', ExlHeader);
  * @param {HTMLHeadElement} headerBlock
  */
 export default async function decorate(headerBlock, options = {}) {
-  if (isFeatureEnabled('isHeaderV2')) {
+  if (isFeatureEnabled('isHeaderV2') || (await isDomainAllowed('headerv2allowedDomains'))) {
     const { default: decorateV2 } = await import('./header-v2.js');
     return decorateV2(headerBlock, options);
   }
